Moved test classes in correct packages.
